Correct rebase errors
